Improved commands and install script

+++ install.sh
@@ -1,4 +1,9 @@
 #!/usr/bin/env bash
+
 sudo ln -s ./quail-compression/run.py /usr/local/bin/qcomp
-echo "Link created"
-rm install.sh
+if [ $? -ne 0 ]; then
+    echo "Install failed!"
+else
+    echo "Link created"
+    rm install.sh
+fi
